requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
facebook-sdk==3.1.0
future==0.18.1
idna==2.8
Pillow==6.2.0
pycparser==2.19
python-telegram-bot==12.2.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
tornado==6.0.3
urllib3==1.25.6
